refactor(decision): extract pure option-scoring engine to lib/decision/option-scoring.ts (C1, BI-E1FB2307)
Decision-surface consolidation, child C1. The pure option-scoring math
(computeStructuredAlignment/computeSemanticAlignment/buildOptionScores/decide
+ types) moves from lib/wiki/principle-decide.ts to a neutral shared module so
the Decision Perspective Gate and the principle_decide MCP handler can share ONE
inner engine instead of two parallel decision paths.

- New: apps/web/lib/decision/option-scoring.ts (verbatim move; engine stays pure,
  zero I/O).
- lib/wiki/principle-decide.ts becomes a back-compat re-export shim — existing
  importers (mcp-tools.ts, kernel/runtime-gate.ts, tests) are unchanged.
- Behavior-preserving: principle-decide tests 26/26 green via the shim; no new
  type errors on the touched files.

This is the inner-engine groundwork; C2 wires the Gate to call it with profile
resolution. BI-E1FB2307.
